# api/cluster/v1/openapi.yaml
...
openapi: 3.0.3
info:
  title: Haliphron Cluster API
  version: 1.0.0
  description: |
    A pull model: the backend never initiates a connection into a cluster.
    The controller registers, leases work through a long poll, acknowledges
    materialization, sends heartbeats and reports the result.

    The semantics of the epoch, lease expiry, report monotonicity and the
    taxonomy of errors are described in `docs/contracts/cluster-api.md`.
    This file is the shape of the messages; that one is the rules.
  contact:
    name: Haliphron
  license:
    name: Proprietary
servers:
  - url: https://{backendHost}/cluster-api/v1
    variables:
      backendHost:
        default: haliphron-backend.haliphron.svc:8082
...
security:
  - clusterJWT: []
paths:
  /register:
    post:
      tags: [registration]
      operationId: registerCluster
      summary: Exchange a one-time bootstrap token for a clusterID
      description: |
        The only call authenticated by a bootstrap token rather than a JWT.

        The controller generates an Ed25519 pair **inside the cluster** and
        transmits only the public half. The private key never leaves the
        cluster, and the backend cannot issue a token on the cluster's behalf.

        Idempotent on the pair (bootstrapToken, publicKey.kid + publicKey.key):
        a repeat with the same key returns the same clusterID and the same
        response, even though the token has already been spent. This saves a
        controller that crashed between a successful response and writing the
        clusterID into its Secret.
      security:
        - bootstrapToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/RegisterRequest'}
      responses:
        '200':
          description: The cluster is registered (or this is a repeat with the same key)
          headers:
            Cache-Control: {$ref: '#/components/headers/NoStore'}
          content:
            application/json:
              schema: {$ref: '#/components/schemas/RegisterResponse'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401':
          description: |
            The bootstrap token is unknown, expired, revoked, or was spent by a
            different public key.
            code=BootstrapTokenInvalid | BootstrapTokenConsumed.
            action=fatal — the operator must issue a new token in the UI.
          content:
            application/problem+json:
              schema: {$ref: '#/components/schemas/Problem'}
        '409':
          description: The cluster name is taken (code=ClusterNameTaken, action=fatal)
          content:
            application/problem+json:
              schema: {$ref: '#/components/schemas/Problem'}
        '422': {$ref: '#/components/responses/UnsupportedController'}
        '429': {$ref: '#/components/responses/RateLimited'}
        '500': {$ref: '#/components/responses/ServerError'}
  /clusters/{clusterID}/leases:
    post:
      tags: [lease]
      operationId: acquireLeases
      summary: Long poll for work
      description: |
        The connection is held until work appears or until `waitSeconds`
        elapses (for the maximum see `maxWaitSeconds` from /register, 30 by
        default — the upper bound is set by the ingress's proxy_read_timeout,
        see docs/architecture.md, section 16).

        The response contains secret material in the clear: the backend cannot
        create a Secret in the cluster, so the controller creates it. The
        response body must not be logged or cached.

        Several controller replicas polling at the same time are safe:
        issuance goes through `FOR UPDATE SKIP LOCKED`, and one lease reaches
        exactly one poller.
      parameters:
        - $ref: '#/components/parameters/ClusterID'
        - $ref: '#/components/parameters/ControllerVersion'
        - $ref: '#/components/parameters/Traceparent'
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/LeaseRequest'}
      responses:
        '200':
          description: One or more leases were issued
          headers:
            Cache-Control: {$ref: '#/components/headers/NoStore'}
          content:
            application/json:
              schema: {$ref: '#/components/schemas/LeaseResponse'}
        '204':
          description: |
            There is no work and the long poll expired normally. The controller
            polls again immediately, without backoff.
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthenticated'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '422': {$ref: '#/components/responses/UnsupportedController'}
        '429': {$ref: '#/components/responses/RateLimited'}
        '500': {$ref: '#/components/responses/ServerError'}
        '503': {$ref: '#/components/responses/Unavailable'}
  /leases/{runID}/ack:
    post:
      tags: [lease]
      operationId: ackLease
      summary: Acknowledge that the lease has been materialized in the cluster
      description: |
        Called after the Secret, the ConfigMap and the `AgentRun` CR have been
        created in the cluster — that is, after the work has become durable and
        will survive a controller restart. The backend moves the run
        `Leased → Dispatched`.

        Before the ack, an `ackDeadline` expiry means the controller crashed
        without managing to create anything and the work definitely did not
        start. The backend raises the epoch and returns the run to `Queued`,
        from where it may go to a different cluster.

        Idempotent on (runID, epoch).
      parameters:
        - $ref: '#/components/parameters/RunID'
        - $ref: '#/components/parameters/ControllerVersion'
        - $ref: '#/components/parameters/Traceparent'
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/AckRequest'}
      responses:
        '200':
          description: Acknowledged (or a repeat — idempotent)
          content:
            application/json:
              schema: {$ref: '#/components/schemas/AckResponse'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthenticated'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/RunNotFound'}
        '409': {$ref: '#/components/responses/EpochMismatch'}
        '429': {$ref: '#/components/responses/RateLimited'}
        '500': {$ref: '#/components/responses/ServerError'}
  /leases/{runID}/artifacts:
    post:
      tags: [lease]
      operationId: mintArtifactBundle
      summary: Issue a fresh presigned bundle for an active lease
      description: |
        The presigned links issued together with the lease live for
        `timeoutSeconds × 2`. A series of infrastructure retries
        (`retry.maxInfraRetries`, 3 by default) can outlast that, and then the
        pod cannot upload the result — a failure that looks like a lost result
        on successful work by the agent.

        Before creating the Job for attempt > 1 the controller is obliged to
        check `expiresAt` and, if necessary, mint a new bundle and update the
        per-run Secret. Requires an active lease with the current epoch.
      parameters:
        - $ref: '#/components/parameters/RunID'
        - $ref: '#/components/parameters/ControllerVersion'
        - $ref: '#/components/parameters/Traceparent'
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/ArtifactBundleRequest'}
      responses:
        '200':
          description: The new bundle
          headers:
            Cache-Control: {$ref: '#/components/headers/NoStore'}
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ArtifactBundle'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthenticated'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/RunNotFound'}
        '409': {$ref: '#/components/responses/EpochMismatch'}
        '429': {$ref: '#/components/responses/RateLimited'}
        '500': {$ref: '#/components/responses/ServerError'}
  /clusters/{clusterID}/heartbeat:
    post:
      tags: [heartbeat]
      operationId: heartbeat
      summary: Liveness, lease renewal, receiving commands
      description: |
        Called every `heartbeatIntervalSeconds` (10 by default). It does four
        things at once:

        1. confirms the cluster is alive (`staleAfterSeconds` without a
           heartbeat → the cluster becomes `Unreachable`);
        2. renews the leases of every run listed;
        3. **reconciles state**: with `reportComplete: true` the `runs` list is
           taken as complete, and the backend can detect work the cluster lost
           immediately rather than waiting for the lease to expire;
        4. delivers commands top-down — the **only channel** in that direction.
           Cancellation is asynchronous, with a delay of one interval.

        The observations in `runs[]` are processed by the same monotonicity
        rules as `/ingest/status`: a heartbeat is a batch of observations plus
        a reconciliation.
      parameters:
        - $ref: '#/components/parameters/ClusterID'
        - $ref: '#/components/parameters/ControllerVersion'
        - $ref: '#/components/parameters/Traceparent'
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/HeartbeatRequest'}
      responses:
        '200':
          description: Accepted
          content:
            application/json:
              schema: {$ref: '#/components/schemas/HeartbeatResponse'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthenticated'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '413': {$ref: '#/components/responses/PayloadTooLarge'}
        '422': {$ref: '#/components/responses/UnsupportedController'}
        '429': {$ref: '#/components/responses/RateLimited'}
        '500': {$ref: '#/components/responses/ServerError'}
        '503': {$ref: '#/components/responses/Unavailable'}
  /ingest/status:
    post:
      tags: [ingest]
      operationId: ingestStatus
      summary: Lifecycle report (batched)
      description: |
        The low-latency path for a phase change. It duplicates the heartbeat's
        observations and obeys the same rules: idempotency on
        (runID, epoch, attempt, phase), monotonicity on (attempt, phaseRank),
        and the first terminal phase wins.

        Batched, because during a reconcile burst the controller receives dozens
        of events per second, and N requests for N events are extra traffic and
        extra chances of a partial failure.

        Reports are applied row by row: partial success is normal. The response
        carries the result for each row; a row with `accepted: false` is not
        repeated by the controller unless `action` is `retry`.
      parameters:
        - $ref: '#/components/parameters/ControllerVersion'
        - $ref: '#/components/parameters/Traceparent'
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/StatusIngestRequest'}
      responses:
        '200':
          description: Processed (see the per-item results)
          content:
            application/json:
              schema: {$ref: '#/components/schemas/StatusIngestResponse'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthenticated'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '413': {$ref: '#/components/responses/PayloadTooLarge'}
        '429': {$ref: '#/components/responses/RateLimited'}
        '500': {$ref: '#/components/responses/ServerError'}
        '503': {$ref: '#/components/responses/Unavailable'}
  /ingest/completion:
    post:
      tags: [ingest]
      operationId: ingestCompletion
      summary: Report on the contents of the result
      description: |
        The forwarding of the pod's webhook (the image contract, section 13.5
        of the architecture), wrapped in an envelope carrying the cluster
        identity and epoch.

        The payload is **references, not contents**, apart from a summary of up
        to 64 KiB. The result is durable in the data plane by the time of the
        call (ADR 15) — on the controller's spool in relay mode, in the object
        store in the other — so this call is an optimization, not a correctness
        condition: if it did not arrive and the terminal status did, the backend
        marks the run `CompletedWithoutResult` and restores the contents by
        reading `runs/{runID}/` back out of the artifact store.

        Idempotent on (runID, attempt).
      parameters:
        - $ref: '#/components/parameters/ControllerVersion'
        - $ref: '#/components/parameters/Traceparent'
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/CompletionIngestRequest'}
      responses:
        '200':
          description: Accepted (or a repeat — idempotent)
          content:
            application/json:
              schema: {$ref: '#/components/schemas/CompletionIngestResponse'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthenticated'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/RunNotFound'}
        '409': {$ref: '#/components/responses/EpochMismatch'}
        '413': {$ref: '#/components/responses/PayloadTooLarge'}
        '429': {$ref: '#/components/responses/RateLimited'}
        '500': {$ref: '#/components/responses/ServerError'}
        '503': {$ref: '#/components/responses/Unavailable'}
...
components:
  securitySchemes:
    clusterJWT:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        A short-lived JWT signed with the **controller's private key**
        (EdDSA / Ed25519). The backend stores only the public half, received at
        registration, and is incapable of issuing a token on the cluster's
        behalf.

        Mandatory claims: `iss` = `sub` = clusterID (ULID),
        `aud` = `haliphron-cluster-api`, `iat`, `exp` (≤ 300 s),
        `jti` (unique). The header carries `kid`.

        Backend checks: the signature by `kid`, `aud`, `exp` with a 60 s clock
        skew tolerance, the cluster's status (`Revoked` → 401), and that `sub`
        matches the `clusterID` in the path (otherwise 403).
    bootstrapToken:
      type: http
      scheme: bearer
      description: |
        A one-time token with a TTL, issued by an operator in the UI. The
        `hlb_` prefix is how the handler tells it from a JWT. It is spent after
        use. Accepted only on `/register`.
  headers:
    NoStore:
      description: The body contains secret material. Do not cache, do not log.
      schema: {type: string, enum: ['no-store']}
...
  schemas:
...
    RegisterRequest:
      type: object
      required: [name, publicKey, controllerVersion, agentNamespace]
      properties:
        name:
          type: string
          maxLength: 63
          pattern: '^[a-z0-9]([-a-z0-9]*[a-z0-9])?$'
          description: A unique human-readable name. Shown in the UI, used in selectors.
          example: prod-eu-west
        labels:
          type: object
          additionalProperties: {type: string}
          description: Labels for a role's `clusterSelector`
          example: {env: prod, region: eu-west-1}
        publicKey: {$ref: '#/components/schemas/PublicKey'}
        controllerVersion: {type: string, example: '1.2.0'}
        agentNamespace:
          type: string
          description: The namespace in which the controller will create agent Jobs
          example: haliphron-agents
        k8sVersion: {type: string, example: v1.31.4}
        runtimes:
          type: array
          description: Which agent runtimes the cluster is prepared to execute. Used in the placement filter.
          items: {$ref: '#/components/schemas/AgentType'}
        capacitySlots:
          type: integer
          format: int32
          minimum: 0
          description: |
            How many concurrent runs the cluster agrees to hold. The upper bound
            is set by the agent namespace's ResourceQuota; this is a value
            declared by the controller, not a computed one.
        crdVersions:
          type: array
          description: The installed versions of the AgentRun CRD. Lets the backend tell whether the cluster is ready for a new spec shape.
          items: {type: string, example: v1alpha1}
...
    RegisterResponse:
      type: object
      description: |
        A dual purpose: confirming the registration and bootstrapping the
        operational parameters. The intervals are set by the control plane
        rather than the cluster's values.yaml — otherwise they drift across
        installations and nobody knows what `staleAfter` actually is.
      required: [clusterID, name, tokenAudience, timings, supportedControllerVersions, serverTime]
      properties:
        clusterID: {$ref: '#/components/schemas/ULID'}
        name: {type: string}
        keyID: {type: string, description: "An echo of `publicKey.kid` — confirmation of which key was accepted"}
        tokenAudience: {type: string, example: haliphron-cluster-api}
        tokenMaxTTLSeconds: {type: integer, format: int32, example: 300}
        timings: {$ref: '#/components/schemas/Timings'}
        supportedControllerVersions:
          type: object
          required: [min, max]
          properties:
            min: {type: string, example: '1.0.0'}
            max: {type: string, example: '1.4.0'}
        serverTime:
          type: string
          format: date-time
          description: |
            For estimating clock skew. A controller whose divergence exceeds the
            JWT tolerance is obliged to log it at startup rather than find out
            through a 401 in production.
...